+++ lib/better_auth_flutter.dart
@@ -64,6 +64,7 @@ export "src/core/api/default/sign_in/models/social/social_provider.dart"
     show SocialProvider;
 export "src/core/api/default/sign_in/models/social/body/sign_in_social_body.dart"
     show SignInSocialBody;
+export "src/core/oauth/oauth_handler.dart" show OAuthHandler, SocialCallback;
 export "src/core/api/models/session/session_response.dart" show SessionResponse;
 export "src/core/api/models/common/sign_out/sign_out_response.dart"
     show SignOutResponse;